requirements.txt
moviepy>=2.0.0
opencv-python>=4.5.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=8.0.0
sentence-transformers>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0